entangled repo split
 * due to the new signature of iota_client_core_init, the iota_client_service folder was deleted.
 * bazel BUILD file was updated
