+++ examples/e02_send_hello.c
@@ -3,8 +3,8 @@
 
 #include "common/trinary/tryte_ascii.h"
 #include <inttypes.h>
-#include "iota_client_service/config.h"
-#include "iota_client_service/client_service.h"
+
+#include "config.h"
 
 static tryte_t const *const ADDRESS =
 (tryte_t *)"ZLGVEQ9JUZZWCZXLWVNTHBDX9G9KZTJP9VEERIIFHY9SIQKYBVAHIMLHXPQVE9IXFDDXNHQINXJDRPFDXNYVAPLZAW";
@@ -56,11 +56,17 @@ retcode_t send_transaction(iota_client_service_t *service){
 }
 
 int main(void){
-    retcode_t ret = RC_ERROR;
-    iota_client_service_t iota_client_service;
-    init_iota_client(&iota_client_service);
+    iota_client_service_t *iota_client_service;
 
-    ret = send_transaction(&iota_client_service);
+#ifdef CONFIG_ENABLE_HTTPS
+    iota_client_service = iota_client_core_init(CONFIG_IRI_NODE_URI, CONFIG_IRI_NODE_PORT, TLS_CERTIFICATE_PEM);
+#else
+    iota_client_service = iota_client_core_init(CONFIG_IRI_NODE_URI, CONFIG_IRI_NODE_PORT, NULL);
+#endif
+
+    retcode_t ret = RC_ERROR;
+    ret = send_transaction(iota_client_service);
+
     if(ret == RC_OK){
         printf("Transaction was sent.\n");
     }else{
