+++ examples/e04_generate_address.c
@@ -4,10 +4,8 @@
 #include "common/trinary/tryte_ascii.h"
 #include "utils/time.h"
 #include <inttypes.h>
-#include "iota_client_service/config.h"
-#include "iota_client_service/client_service.h"
 
-#include "iota_client_service/config.h"
+#include "config.h"
 
 static tryte_t const *const SEED =
         (tryte_t *)"G9JEMIRJKUXDKUPPAIMEQSGVADYLSJRSBTEIRDWSCTLCVQOJWBM9XESTWTSONOTDDQUXMYCNVAKZWPPYW";
@@ -40,11 +38,17 @@ retcode_t get_new_address(iota_client_service_t *s) {
 }
 
 int main(void){
-    retcode_t ret = RC_ERROR;
-    iota_client_service_t iota_client_service;
-    init_iota_client(&iota_client_service);
+    iota_client_service_t *iota_client_service;
 
-    ret = get_new_address(&iota_client_service);
+#ifdef CONFIG_ENABLE_HTTPS
+    iota_client_service = iota_client_core_init(CONFIG_IRI_NODE_URI, CONFIG_IRI_NODE_PORT, TLS_CERTIFICATE_PEM);
+#else
+    iota_client_service = iota_client_core_init(CONFIG_IRI_NODE_URI, CONFIG_IRI_NODE_PORT, NULL);
+#endif
+
+    retcode_t ret = RC_ERROR;
+    ret = get_new_address(iota_client_service);
+
     if(ret == RC_OK){
         printf("Addresses generated.\n");
     }else{
