+++ examples/e05_check_balance.c
@@ -5,9 +5,7 @@
 #include "utils/time.h"
 #include <inttypes.h>
 
-#include "iota_client_service/config.h"
-#include "iota_client_service/client_service.h"
-
+#include "config.h"
 
 static tryte_t const *const ADDRESS =
         (tryte_t *)"TOKLOARHKXQCVPPVVIPIJGLUTLTKFHYGMBBLOXJFYGSARLOTYFFSDZNYCOBOCNPGRMJWZCQBNOROUCE9G";
@@ -65,11 +63,17 @@ retcode_t get_balance(iota_client_service_t *service) {
 }
 
 int main(void){
-    retcode_t ret = RC_ERROR;
-    iota_client_service_t iota_client_service;
-    init_iota_client(&iota_client_service);
+    iota_client_service_t *iota_client_service;
 
-    ret = get_balance(&iota_client_service);
+#ifdef CONFIG_ENABLE_HTTPS
+    iota_client_service = iota_client_core_init(CONFIG_IRI_NODE_URI, CONFIG_IRI_NODE_PORT, TLS_CERTIFICATE_PEM);
+#else
+    iota_client_service = iota_client_core_init(CONFIG_IRI_NODE_URI, CONFIG_IRI_NODE_PORT, NULL);
+#endif
+
+    retcode_t ret = RC_ERROR;
+    ret = get_balance(iota_client_service);
+
     if(ret == RC_OK){
         printf("Check balances done\n");
     }else{
