--- iota_client_service/BUILD
@@ -1,15 +1,0 @@
-package(default_visibility = ["//visibility:public"])
-
-cc_library(
-    name = "service",
-    srcs = [
-        "client_service.c",
-    ],
-    hdrs = [
-        "client_service.h",
-        "config.h"
-    ],
-    deps = [
-        "@entangled//cclient/api",
-    ],
-)
