--- iota_client_service/client_service.h
@@ -1,9 +1,0 @@
-#ifndef C_IOTA_WORKSHOP_CLIENT_SERVICE_H
-#define C_IOTA_WORKSHOP_CLIENT_SERVICE_H
-
-#include "cclient/api/core/core_api.h"
-#include "cclient/api/extended/extended_api.h"
-
-void init_iota_client(iota_client_service_t *const service);
-
-#endif //C_IOTA_WORKSHOP_CLIENT_SERVICE_H
